# OilCaseX/docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OilCase Api",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/Purchased/Seismic": {
      "get": {
        "tags": [
          "Seismic"
        ],
        "summary": "Возвращает купленные сейсмики",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PurchasedSeismic"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Seismic"
        ],
        "summary": "Добавляет купленный объект",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchasedSeismic"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchasedSeismic"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PurchasedSeismic"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Id нового объекта",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Уже есть купленная сейсмика с данными координатами",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Purchased/Seismic/{id}": {
      "delete": {
        "tags": [
          "Seismic"
        ],
        "summary": "Удаление сейсмики",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Указаный Id прнеадлежит другой команде или шаг покупки уже завершен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/Purchased/WellTop": {
      "post": {
        "tags": [
          "WellTop"
        ],
        "summary": "Добавление отбивок",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WellTop"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/WellTop"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/WellTop"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Скважина для отбивки не найдена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/Purchased/WellTop/{id}": {
      "delete": {
        "tags": [
          "WellTop"
        ],
        "summary": "Удаление отбивок скважин",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "PurchasedSeismic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "gameStep": {
            "type": "integer",
            "format": "int32"
          },
          "startCellX": {
            "type": "integer",
            "format": "int32"
          },
          "startCellY": {
            "type": "integer",
            "format": "int32"
          },
          "endCellX": {
            "type": "integer",
            "format": "int32"
          },
          "endCellY": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "WellTop": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "purchasedBoreholeId": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "z": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      }
    }
  }
}